+++ lib/beta-completion/index.ts
@@ -1,0 +1,1 @@
+export * from "./beta31Checklist";